// packages/app/src/auth/enum/pin-setup-reducer-action.enum.ts
export enum PinSetupReducerActionEnum {
    ADD_DIGIT = 'ADD_DIGIT',
    DELETE_DIGIT = 'DELETE_DIGIT',
    RESET_INPUT = 'RESET_INPUT',
    SET_ERROR = 'SET_ERROR',
    SET_LOADING = 'SET_LOADING',
    VERIFY_OLD_SUCCESS = 'VERIFY_OLD_SUCCESS',
    STORE_NEW_PIN = 'STORE_NEW_PIN',
    NEXT_STEP = 'NEXT_STEP',
    RESET = 'RESET',
    CLEAR_JUST_COMPLETED = 'CLEAR_JUST_COMPLETED'
}